Refactor table to object mapping into dedicated classes.
Each table now has its own file under the 'table' module. This prepares
to restructure the database schema.
